Add GUI functionality to the Yong chatbot
