--- src/main/java/duke/exception/DukeException.java
@@ -1,10 +1,0 @@
-package duke.exception;
-
-/**
- * Class to handle exceptions due to incorrect user input.
- */
-public class DukeException extends RuntimeException {
-    public DukeException(String message) {
-        super(message);
-    }
-}
