+++ src/main/java/yong/command/DeleteCommand.java
@@ -1,0 +1,41 @@
+package yong.command;
+
+import yong.tasklist.TaskList;
+
+import yong.tasks.Task;
+
+/**
+ * Class representing actions run when the delete command is called by user.
+ */
+public class DeleteCommand extends Command {
+
+    String numberString;
+
+    /**
+     * Constructor for DeleteCommand
+     *
+     * @param taskList TaskList object being used to maintain the list of tasks in the chatbot.
+     * @param numberString Parsed string input from the CLI, represents task to be deleted.
+     */
+    public DeleteCommand(TaskList taskList, String numberString) {
+        super(taskList);
+        this.numberString = numberString;
+
+    }
+
+    /**
+     * Method to be executed when this command is called.
+     * Deletes the specified task by the user.
+     */
+    @Override
+    public String execute() {
+        Integer number = Integer.parseInt(numberString);
+
+        Task task = taskList.delete(number);
+
+        outputString = "YONG has deleted this task for you! \n" + task.toString();
+
+        return outputString;
+
+    }
+}
